LayoutTests/inspector/unit-tests/mimetype-utilities-expected.txt


== Running test suite: MIMETypeUtilities
-- Running test case: fileExtensionForURL
PASS: File extension for null URL should be null.
PASS: File extension for invalid URL should be null.
PASS: File extension for URL without last path component should be null.
PASS: File extension for URL without last path component should be null.
PASS: File extension for URL with last path component without a period should be null.
PASS: File extension for URL with last path component without a period should be null.
PASS: File extension for URL with last path component ending in a period should be null.
PASS: File extension for "foo.xyz" should be "xyz".
PASS: File extension for "image.png" should be "png".
PASS: File extension for "image.png" should be "gif".
PASS: File extension for "script.js" should be "js".
PASS: File extension for "script.min.js" should be "js".

-- Running test case: fileExtensionForMIMEType
PASS: File extension for null mime type should be null.
PASS: File extension for invalid mime type should be null.
PASS: File extension for unknown mime type should be null.
PASS: File extension for "image/jpeg" should be "jpg".
PASS: File extension for "image/png" should be "png".
PASS: File extension for "image/gif" should be "gif".
PASS: File extension for "text/javascript" should be "js".
PASS: File extension for "application/json" should be "json".
PASS: File extension for "application/vnd.api+json" should be "json".
PASS: File extension for "application/xhtml+xml" should be "xhtml".
PASS: File extension for "image/svg+xml" should be "svg".
PASS: File extension for "text/foo+xml" should be "xml".
